Problem: using deprecated function in downloading (fix #356) Solution: Fix connection and request failures in PlayCppSdkDownloader

+++ Source/CronosPlayUnreal/Public/PlayCppSdkDownloader.h
@@ -67,7 +67,7 @@ class CRONOSPLAYUNREAL_API UPlayCppSdkDownloader : public UObject {
     UPROPERTY(BlueprintAssignable)
     FPlayDownloadProgress PlayDownloadProgressCallback;
     void OnPlayDownloadProgressCallback(FHttpRequestPtr RequestPtr,
-                                        int32 SendBytes, int32 RecvBytes);
+                                        uint64 SendBytes, uint64 RecvBytes);
 
     /**
      * SetupIpfsConverting
